sentiment-terminal/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
openpyxl
python-dotenv
httpx
transformers
torch
pyarrow
fastparquet
groq
requests
feedparser
textstat
scipy
scikit-learn
beautifulsoup4
